requirements.txt
fastapi[standard]
alembic
python-jose[cryptography]
passlib[bcrypt]
aiomysql
sqlalchemy[asyncio]